MNT: Move Command helper class to testing/utils
